refactor: tsconfig에 baseUrl, paths 적용 (2)
