+++ src/components/Todo/TodoList.js
@@ -1,5 +1,5 @@
 import {Component} from "@/core";
-import {DELETE_ITEM, SET_EDITING, TOGGLE_ITEM, UPDATE_ITEM, UPDATE_ITEM_PRIORITY, todoOfTeamStore} from "../../store/todoOfTeamStore";
+import {DELETE_ITEM, SET_EDITING, TOGGLE_ITEM, UPDATE_ITEM, UPDATE_ITEM_PRIORITY, todoOfTeamStore} from "@/store/todoOfTeamStore";
 import {TodoListFooter} from "./TodoListFooter";
 import {TodoItemAppender} from "./TodoItemAppender";
 import { PriorityTypes, FilterTypes } from "@/constants";
