+++ src/services/index.ts
@@ -1,0 +1,2 @@
+export { default as TeamService } from './TeamService';
+export { default as TodoService } from './TodoService';
