+++ src/store/teamStore.js
@@ -1,5 +1,5 @@
 import {Store} from "@/core";
-import TeamService from "@/services/TeamService";
+import {TeamService} from "@/services";
 
 export const SET_TEAMS = 'SET_TEAMS';
 export const SET_OPENED_APPEND_FORM = 'SET_OPENED_APPEND_FORM';
